Fix compilation warnings: type conversions, shadow variables, unused parameters

+++ src/Bag.cpp
@@ -6,8 +6,8 @@ int Bag::next() {
     if (bag.empty()) {
         bag = {0,1,2,3,4,5,6};
 
-        for (int i = bag.size() - 1; i > 0; i--) {
-            int j = rand() % (i + 1);
+        for (size_t i = bag.size() - 1; i > 0; i--) {
+            size_t j = static_cast<size_t>(rand()) % (i + 1);
             std::swap(bag[i], bag[j]);
         }
     }
